Implemented Request Validator.

+++ trunk/cofm/server/src/collab/data/bean/Votable.java
@@ -7,8 +7,8 @@ import java.util.TreeSet;
 public class Votable<T> {
 	
 	protected T value;
-	protected SortedSet<Long> support = new TreeSet<Long>(); // List of User ID
-	protected SortedSet<Long> against = new TreeSet<Long>(); // List of User ID
+	protected SortedSet<Integer> support = new TreeSet<Integer>(); // List of User ID
+	protected SortedSet<Integer> against = new TreeSet<Integer>(); // List of User ID
 	
 	public Votable() {
 		
@@ -26,18 +26,22 @@ public class Votable<T> {
 		this.value = value;
 	}
 	
-	public synchronized void voteYes(Long userid) {
+	public synchronized void voteYes(Integer userid) {
 		// An user either support or against the value
 		support.add(userid);
 		against.remove(userid);
 	}
 	
-	public synchronized void voteNo(Long userid) {
+	public synchronized void voteNo(Integer userid) {
 		support.remove(userid);
 		against.add(userid);
 	}
 	
-	public void support(T val, Long userid) {
+	// NOTES: For a group of candidates, say {A, B, C}. 
+	//  Support A means against B and C; that is, support one means against others.
+	//  But against one does NOT mean support others, maybe you want to against them all; that's
+	//  why the against() method has no "else" part in it.
+	public void support(T val, Integer userid) {
 		if (val.equals(value)) {
 			voteYes(userid);
 		} else {
@@ -45,9 +49,10 @@ public class Votable<T> {
 		}
 	}
 	
-	public void against(T val, Long userid) {
+	public void against(T val, Integer userid) {
 		if (val.equals(value)) {
 			voteNo(userid);
 		} 
+		// NO else part here. See the "NOTES" above.
 	}
 }
